requirements.txt
asgiref==3.3.1
Django==3.1.3
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
phonenumbers==8.12.19
phonenumberslite==8.12.19
pytz==2021.1
sqlparse==0.4.1
